pyproject.toml
[project]
name = "cinephix-backend"
version = "0.1.0"
description = "Backend FastAPI para CinePhix SaaS"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Adrián Vázquez", email = "adrian@vazquezdev.pro"}
]

dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy[asyncio]==2.0.35",
    "asyncpg==0.29.0",
    "alembic==1.13.2",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "python-jose[cryptography]==3.3.0",
    "passlib[argon2]==1.7.4",
    "bcrypt==4.2.0",
    "redis[hiredis]==5.0.8",
    "httpx==0.27.2",
    "resend==2.0.0",
    "python-multipart==0.0.12",
    "email-validator==2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "httpx==0.27.2",
    "ruff==0.6.9",
    "mypy==1.11.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true